Details View complete
I’m showing he details of a particular contact. And hiding anything
else on the screen while we’re doing that. So if we’re viewing details
and click ‘add contact’ the details disappear and the form shows. And
the opposite if we’re adding a contact then viewing the details